requirements.txt
mimesis==17.0.0
dnspython>=2.1.0
pymongo>=4.6.2
locust==2.24.0
flask>=2.0,<3.0
Click==7.1.2
deepmerge==0.0.5
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.2
Werkzeug==2.0.3
wheel==0.37.1
python-dotenv>=1.0.0
